backend/requirements.txt
fastapi
uvicorn
pydantic[email]
axios
weasyprint
jinja2
python-dotenv